Rolling back to revision 1285. Drat these computers, they're so naughty and so complex.

+++ branches/sketches/net/roundTripTimer.C
@@ -47,7 +47,7 @@ int32_t TupleTimestamp::delay()
   boost::posix_time::ptime  now;
   getTime(now);
 
-  return((int32_t)(now - tt_).total_milliseconds());
+  return((now - tt_).total_milliseconds());
 }
 
 /////////////////////////////////////////////////////////////////////
